remove glide.lock and glide.yaml - dropping glide support
glide has been not been maintained. It's website has gone offline.
glide was old fashioned when dep was hot more than two or three years
ago. The go ecosystem has moved on to go modules. In a near-future
version of go building with modules may no longer be supported [1].

So, despite our earnest desire to stop maintaining heketi it still lives
on and has numerous users in the field. Thus we make this step the start
of our switch-over to go modules by eliminating our dependency on glide.

[1] - https://blog.golang.org/go116-module-changes#TOC_2.

Signed-off-by: John Mulligan <jmulligan@redhat.com>